+++ module/apps/roll-dialog.js
@@ -1,5 +1,6 @@
 /* global Dialog, FormData, game, renderTemplate */
 
+import { CoC7ChatMessage } from './coc7-chat-message.js'
 import { CoC7Check } from '../check.js'
 
 export class RollDialog {
@@ -19,24 +20,27 @@ export class RollDialog {
         options.difficultyLevel.extreme = true
       }
     }
-
     // if( undefined == options.askValue) options.askValue = true;
     if (options.name && !options.displayName) options.displayName = options.name
     const unknownDifficultyDefault =
       game.settings.get('CoC7', 'defaultCheckDifficulty') === 'unknown'
+    const data = {
+      allowFlatDiceModifier:
+        game.settings.get('CoC7', 'allowFlatDiceModifier') &&
+        !options.disableFlatDiceModifier,
+      allowFlatThresholdModifier:
+        game.settings.get('CoC7', 'allowFlatThresholdModifier') &&
+        !options.disableFlatThresholdModifier,
+      difficulty: CoC7Check.difficultyLevel,
+      unknownDifficultyDefault: unknownDifficultyDefault,
+      options
+    }
+    if (typeof options.cardType !== 'undefined') {
+      data.cardTypes = CoC7ChatMessage.cardTypes(options)
+    }
     const html = await renderTemplate(
       'systems/CoC7/templates/apps/bonus.html',
-      {
-        allowFlatDiceModifier:
-          game.settings.get('CoC7', 'allowFlatDiceModifier') &&
-          !options.disableFlatDiceModifier,
-        allowFlatThresholdModifier:
-          game.settings.get('CoC7', 'allowFlatThresholdModifier') &&
-          !options.disableFlatThresholdModifier,
-        difficulty: CoC7Check.difficultyLevel,
-        unknownDifficultyDefault: unknownDifficultyDefault,
-        options
-      }
+      data
     )
     return new Promise(resolve => {
       let formData = null