+++ module/chat/cards/combined-roll.js
@@ -69,10 +69,17 @@ export class CombinedCheckCard extends RollCard {
     return !this.success
   }
 
+  get checkGMInitiator () {
+    return game.users.get(this.initiator)?.isGM
+  }
+
   async getHtmlRoll () {
     if (!this.rolled) return undefined
     const check = new CoC7Check()
-    await check._perform({ roll: this._roll[check.diceModifier || 0], silent: true })
+    await check._perform({
+      roll: this._roll[check.diceModifier || 0],
+      silent: true
+    })
     return await check.getHtmlRoll({ hideSuccess: true })
   }
 
@@ -184,6 +191,11 @@ export class CombinedCheckCard extends RollCard {
 
   async compute () {
     if (!this._roll) return
+
+    this.rolls = this.rolls.filter(roll => {
+      return typeof roll.actor.data !== 'undefined' // remove any actors that no longer exist
+    })
+
     for (const r of this.rolls) {
       if (!r.rolled) {
         r.modifier = r.diceModifier || 0
@@ -200,10 +212,6 @@ export class CombinedCheckCard extends RollCard {
         })
       }
     }
-
-    this.rolls = this.rolls.filter(roll => {
-      return typeof roll.actor.data !== 'undefined' // Check if there's an actor set and if there's one and it doesnt exist remove him.
-    })
 
     this._htmlRoll = await this.getHtmlRoll()
   }